--- src/automaton/builders.rs
@@ -1,3 +1,0 @@
-mod line_builder;
-
-pub use line_builder::LineBuilder;
