--- src/new_api_v04.rs
@@ -1,9 +1,0 @@
-//! The new API for portmatching v0.4.
-//!
-//! This module is just a temporary space to develop the new API.
-//! This will replace the current API as soon as we agree on a stable interface.
-
-pub mod constraint;
-pub mod mutex_tree;
-pub mod predicate;
-pub mod variable;
