--- src/patterns.rs
@@ -1,7 +1,0 @@
-mod line_pattern;
-mod pattern;
-
-pub use line_pattern::LinePattern;
-pub(crate) use line_pattern::{IterationStatus, Line, PredicatesIter};
-pub(crate) use pattern::{compatible_offsets, UnweightedEdge};
-pub use pattern::{Edge, NoRootFound, Pattern, UnweightedPattern, WeightedPattern};
